Rename package keyboardmanager to keyboard
