+++ keyboard/src/androidLibMain/AndroidManifest.xml
@@ -1,0 +1,3 @@
+<manifest package="com.splendo.kaluga.keyboard"
+          xmlns:android="http://schemas.android.com/apk/res/android"
+/>
