--- keyboardmanager/src/androidLibAndroidTest/AndroidManifest.xml
@@ -1,6 +1,0 @@
-<manifest package="com.splendo.kaluga.keyboardmanager"
-          xmlns:android="http://schemas.android.com/apk/res/android">
-    <application android:theme="@style/Theme.AppCompat">
-        <activity android:name=".TestActivity"/>
-    </application>
-</manifest>
