--- keyboardmanager/src/androidLibAndroidTest/kotlin/TestActivity.kt
@@ -1,20 +1,0 @@
-package com.splendo.kaluga.keyboardmanager
-
-import android.os.Bundle
-import android.widget.LinearLayout
-import android.widget.TextView
-import androidx.appcompat.app.AppCompatActivity
-
-class TestActivity: AppCompatActivity() {
-
-    lateinit var textView: TextView
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        val view = LinearLayout(this)
-        textView = TextView(this)
-        view.addView(textView)
-        setContentView(view)
-    }
-}
