--- keyboardmanager/src/androidLibMain/AndroidManifest.xml
@@ -1,3 +1,0 @@
-<manifest package="com.splendo.kaluga.keyboardmanager"
-          xmlns:android="http://schemas.android.com/apk/res/android"
-/>
